+++ app/src/main/java/org/simple/clinic/facility/alertchange/AlertFacilityChangeSheet.kt
@@ -1,29 +1,45 @@
 package org.simple.clinic.facility.alertchange
 
-import android.app.Activity
+import android.app.Dialog
 import android.content.Context
 import android.content.Intent
-import android.content.res.Configuration
 import android.os.Bundle
 import android.os.Parcelable
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import com.f2prateek.rx.preferences2.Preference
-import io.github.inflationx.viewpump.ViewPumpContextWrapper
 import kotlinx.android.parcel.Parcelize
-import org.simple.clinic.ClinicApp
 import org.simple.clinic.R
 import org.simple.clinic.databinding.SheetAlertFacilityChangeBinding
-import org.simple.clinic.facility.change.FacilityChangeActivity
+import org.simple.clinic.di.injector
+import org.simple.clinic.facility.alertchange.Continuation.ContinueToActivity
+import org.simple.clinic.facility.alertchange.Continuation.ContinueToScreen
+import org.simple.clinic.facility.change.FacilityChangeScreen
 import org.simple.clinic.feature.Features
+import org.simple.clinic.mobius.ViewRenderer
+import org.simple.clinic.navigation.v2.ExpectsResult
+import org.simple.clinic.navigation.v2.Router
+import org.simple.clinic.navigation.v2.ScreenKey
+import org.simple.clinic.navigation.v2.ScreenResult
+import org.simple.clinic.navigation.v2.Succeeded
+import org.simple.clinic.navigation.v2.compat.wrap
+import org.simple.clinic.navigation.v2.fragments.BaseBottomSheet
+import org.simple.clinic.router.ScreenResultBus
 import org.simple.clinic.router.screen.FullScreenKey
-import org.simple.clinic.util.unsafeLazy
-import org.simple.clinic.util.withLocale
-import org.simple.clinic.util.wrap
-import org.simple.clinic.widgets.BottomSheetActivity
+import org.simple.clinic.router.util.resolveFloat
 import java.util.Locale
 import javax.inject.Inject
 import javax.inject.Named
 
-class AlertFacilityChangeSheet : BottomSheetActivity() {
+class AlertFacilityChangeSheet :
+    BaseBottomSheet<
+        AlertFacilityChangeSheet.Key,
+        SheetAlertFacilityChangeBinding,
+        AlertFacilityChangeModel,
+        AlertFacilityChangeEvent,
+        AlertFacilityChangeEffect>(),
+    ExpectsResult {
 
   @Inject
   lateinit var locale: Locale
@@ -35,40 +51,30 @@ class AlertFacilityChangeSheet : BottomSheetActivity() {
   @Inject
   lateinit var features: Features
 
-  private lateinit var component: AlertFacilityChangeComponent
+  @Inject
+  lateinit var screenResults: ScreenResultBus
 
-  companion object {
-    const val FACILITY_CHANGE = 101
-    private const val CURRENT_FACILITY_NAME = "current_facility"
-    private const val CONTINUE_TO = "continue_to"
+  @Inject
+  lateinit var router: Router
 
-    private const val EXTRA_CONTINUE_TO = "extra_continue_to"
+  override fun defaultModel() = AlertFacilityChangeModel()
 
-    fun intent(
-        context: Context,
-        currentFacilityName: String,
-        continuation: Continuation
-    ): Intent {
-      val intent = Intent(context, AlertFacilityChangeSheet::class.java)
-      intent.putExtra(CURRENT_FACILITY_NAME, currentFacilityName)
-      intent.putExtra(CONTINUE_TO, continuation)
-      return intent
-    }
-
-    fun <T : Parcelable> readContinuationExtra(intent: Intent): T {
-      return intent.getParcelableExtra<T>(EXTRA_CONTINUE_TO)!!
+  override fun uiRenderer(): ViewRenderer<AlertFacilityChangeModel> {
+    return object : ViewRenderer<AlertFacilityChangeModel> {
+      override fun render(model: AlertFacilityChangeModel) {
+        // Nothing to render here
+      }
     }
   }
 
-  private val currentFacilityName by unsafeLazy {
-    intent.getStringExtra(CURRENT_FACILITY_NAME)!!
-  }
+  override fun bindView(inflater: LayoutInflater, container: ViewGroup?) =
+      SheetAlertFacilityChangeBinding.inflate(inflater, container, false)
 
-  private val continuation by unsafeLazy {
-    intent.getParcelableExtra<Continuation>(CONTINUE_TO)!!
-  }
+  private val currentFacilityName
+    get() = screenKey.currentFacilityName
 
-  private lateinit var binding: SheetAlertFacilityChangeBinding
+  private val continuation
+    get() = screenKey.continuation
 
   private val facilityName
     get() = binding.facilityName
@@ -79,18 +85,28 @@ class AlertFacilityChangeSheet : BottomSheetActivity() {
   private val changeButton
     get() = binding.changeButton
 
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
+  override fun onAttach(context: Context) {
+    super.onAttach(context)
+    context.injector<Injector>().inject(this)
+  }
+
+  override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
+    return super.onCreateDialog(savedInstanceState).apply {
+      window!!.setDimAmount(0F)
+    }
+  }
+
+  override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+    super.onViewCreated(view, savedInstanceState)
 
     if (isFacilitySwitchedPreference.get().not()) {
-      closeSheetWithContinuation()
+      view.postDelayed(::closeSheetWithContinuation, 100)
     } else {
-      binding = SheetAlertFacilityChangeBinding.inflate(layoutInflater)
-      setContentView(binding.root)
+      showDialogUi()
 
       facilityName.text = getString(R.string.alertfacilitychange_facility_name, currentFacilityName)
       yesButton.setOnClickListener {
-        closeSheetWithResult(Activity.RESULT_OK)
+        proceedToNextScreen()
       }
 
       changeButton.setOnClickListener {
@@ -99,57 +115,62 @@ class AlertFacilityChangeSheet : BottomSheetActivity() {
     }
   }
 
-  override fun attachBaseContext(baseContext: Context) {
-    setupDI()
-
-    val wrappedContext = baseContext
-        .wrap { ViewPumpContextWrapper.wrap(it) }
-
-    super.attachBaseContext(wrappedContext)
-    applyOverrideConfiguration(Configuration())
+  private fun showDialogUi() {
+    val backgroundDimAmount = requireContext().resolveFloat(android.R.attr.backgroundDimAmount)
+    requireDialog().window!!.setDimAmount(backgroundDimAmount)
+    binding.root.visibility = View.VISIBLE
   }
 
-  override fun applyOverrideConfiguration(overrideConfiguration: Configuration) {
-    super.applyOverrideConfiguration(overrideConfiguration.withLocale(locale, features))
-  }
-
-  private fun setupDI() {
-    component = ClinicApp.appComponent
-        .alertFacilityChangeComponent()
-        .create(activity = this)
-
-    component.inject(this)
-  }
-
-  override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-    super.onActivityResult(requestCode, resultCode, data)
-
-    if (requestCode == FACILITY_CHANGE) {
-      closeSheetWithResult(resultCode)
+  override fun onScreenResult(requestType: Parcelable, result: ScreenResult) {
+    if (requestType == ChangeCurrentFacility && result is Succeeded) {
+      proceedToNextScreen()
     }
   }
 
-  private fun closeSheetWithResult(resultCode: Int) {
-    if (resultCode == Activity.RESULT_OK) {
-      isFacilitySwitchedPreference.set(false)
-      closeSheetWithContinuation()
-    } else {
-      val intent = Intent()
-      setResult(resultCode, intent)
-      finish()
-    }
+  private fun proceedToNextScreen() {
+    isFacilitySwitchedPreference.set(false)
+    closeSheetWithContinuation()
   }
 
   private fun closeSheetWithContinuation() {
-    val intent = Intent()
-    intent.putExtra(EXTRA_CONTINUE_TO, continuation)
-    setResult(Activity.RESULT_OK, intent)
-    finish()
+    when (continuation) {
+      is ContinueToScreen -> {
+        val screenKey = (continuation as ContinueToScreen).screenKey
+        router.replaceTop(screenKey.wrap())
+      }
+      is ContinueToActivity -> {
+        val (intent, requestCode) = (continuation as ContinueToActivity).run {
+          intent to requestCode
+        }
+        requireActivity().startActivityForResult(intent, requestCode)
+        router.pop()
+      }
+    }
   }
 
   private fun openFacilityChangeScreen() {
-    startActivityForResult(FacilityChangeActivity.intent(this), FACILITY_CHANGE)
+    router.pushExpectingResult(ChangeCurrentFacility, FacilityChangeScreen.Key())
   }
+
+  @Parcelize
+  data class Key(
+      val currentFacilityName: String,
+      val continuation: Continuation
+  ) : ScreenKey() {
+
+    override val analyticsName = "Alert Facility Change"
+
+    override val type = ScreenType.Modal
+
+    override fun instantiateFragment() = AlertFacilityChangeSheet()
+  }
+
+  interface Injector {
+    fun inject(target: AlertFacilityChangeSheet)
+  }
+
+  @Parcelize
+  object ChangeCurrentFacility : Parcelable
 }
 
 sealed class Continuation : Parcelable {